v2.5.0 release preparation
- Bump version 2.4.0 → 2.5.0 (Cargo.toml + pyproject.toml)
- Fix all clippy warnings (207 auto-fixed + 48 manual, 0 errors remaining)
- Fix CalibrationStats Default impl (min=0.0 → INFINITY)
- Standardize CI: cargo fmt --check + cargo clippy --all-targets -- -D warnings
- Remove orphaned baseline scripts (scripts/ci/)
- Remove temp files (plan.md, debug_bench3.py, profile_each_layer.py)
- Add CHANGELOG.md v2.5.0 entry
- Fix 188 SAFETY annotations, run cargo fmt on entire codebase

+++ src/python/tensor.rs
@@ -434,9 +434,9 @@ impl PyTensor {
         let (start, stop, step) = if let Ok(slice) = idx.cast::<PySlice>() {
             let indices = slice.indices(dim0)?;
             (
-                indices.start as isize,
-                indices.stop as isize,
-                indices.step as isize,
+                indices.start,
+                indices.stop,
+                indices.step,
             )
         } else {
             let idx_val: isize = idx.extract()?;
